Get rid of `JMSParserLegacy` file/tests
When `readonly` became a reserved word in PHP 8.1, it was required to have two implementations of the `JMSParser`. One for versions below PHP 8.0, and one for PHP 8.1 and higher.
However, the jms serializer library itself now renamed the offending class to `DeprecatedReadOnly`  (with an alias to the old clas name) and it also extends the `ReadOnlyProperty` attribute, so having a separate class for it in here is no longer necessary.
